fold ray utils into existing utils.py
