+++ modin/pandas/test/test_groupby.py
@@ -1974,25 +1974,27 @@ def test_validate_by():
 
 def test_groupby_with_virtual_partitions():
     # from https://github.com/modin-project/modin/issues/4464
-    md_df, pd_df = create_test_dfs(test_data["int_data"])
+    modin_df, pandas_df = create_test_dfs(test_data["int_data"])
 
     # Concatenate DataFrames here to make virtual partitions.
-    big_md_df = pd.concat([md_df for _ in range(5)])
-    big_pd_df = pandas.concat([pd_df for _ in range(5)])
+    big_modin_df = pd.concat([modin_df for _ in range(5)])
+    big_pandas_df = pandas.concat([pandas_df for _ in range(5)])
 
     # Check that the constructed Modin DataFrame has virtual partitions when
     # using Ray, and doesn't when using another execution engines.
     if Engine.get() == "Ray":
         assert issubclass(
-            type(big_md_df._query_compiler._modin_frame._partitions[0][0]),
+            type(big_modin_df._query_compiler._modin_frame._partitions[0][0]),
             PandasDataframeAxisPartition,
         )
     else:
         assert not issubclass(
-            type(big_md_df._query_compiler._modin_frame._partitions[0][0]),
+            type(big_modin_df._query_compiler._modin_frame._partitions[0][0]),
             PandasDataframeAxisPartition,
         )
-    eval_general(big_md_df, big_pd_df, lambda df: df.groupby(df.columns[0]).count())
+    eval_general(
+        big_modin_df, big_pandas_df, lambda df: df.groupby(df.columns[0]).count()
+    )
 
 
 @pytest.mark.parametrize("sort", [True, False])
